requirements.txt
aiofiles==25.1.0
aiohttp==3.13.3
aioodbc==0.5.0
asyncpg==0.31.0
atproto==0.0.65
azure-containerregistry==1.2.0
azure-identity==1.25.3
azure-storage-blob==12.28.0
colorama==0.4.6
discord.py==2.7.1
fastapi==0.135.1
gunicorn==25.1.0
lumaai==1.20.1
mcp==1.26.0
openai==2.28.0
pyodbc==5.3.0
python-dotenv==1.2.2
python-jose==3.5.0
tiktoken==0.12.0
uvicorn==0.41.0
